pyproject.toml
[tool.poetry]
name = "wafiqkicks-be"
version = "0.1.0"
description = "Baby kicking counter - Backend"
authors = ["Syafiq <ahmadsyafiq93@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.4"
djangorestframework = "^3.12.2"
django-cors-headers = "^3.6.0"
gunicorn = "^20.0.4"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
ipython = "^7.21.0"
isort = "^5.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
include_trailing_comma = "True"
force_grid_wrap = 0
use_parentheses = "True"
ensure_newline_before_comments = "True"
line_length = 88

[tool.pylint.messages_control]
disable = "C0330, C0326"

[tool.pylint.format]
max-line-length = "88"
